LayoutTests/platform/ios/ios/plugin/youtube-flash-plugin-iframe-expected.txt

Blocked access to external URL http://www.youtube.com/embed/N0gb9v4LI4o
Blocked access to external URL http://www.youtube.com/embed/N0gb9v4LI4o
Blocked access to external URL http://www.youtube.com/embed/N0gb9v4LI4o?version=3&hl=en_US
Blocked access to external URL http://www.youtube.com/embed/N0gb9v4LI4o?version=3&hl=en_US
Blocked access to external URL http://www.youtube.com/embed/N0gb9v4LI4o
Test transforming youtube flash plugin to youtube iFrame based player. < rdar://problem/11201356>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS normalEmbed.tagName is "EMBED"
PASS elinkEmbed.tagName is "EMBED"
PASS objectEmbed.tagName is "EMBED"
PASS objectNoEmbed.tagName is "OBJECT"
PASS document.querySelectorAll("iframe").length is 1
PASS internals.shadowPseudoId(normalEmbedShadowRoot.firstChild) is "-webkit-plugin-replacement"
PASS normalEmbedShadowRoot.firstChild.firstChild.tagName is "IFRAME"
PASS internals.shadowPseudoId(objectEmbedShadowRoot.firstChild) is "-webkit-plugin-replacement"
PASS objectEmbedShadowRoot.firstChild.firstChild.tagName is "IFRAME"
PASS internals.shadowPseudoId(objectNoEmbedShadowRoot.firstChild) is "-webkit-plugin-replacement"
PASS objectNoEmbedShadowRoot.firstChild.firstChild.tagName is "IFRAME"
PASS internals.shadowPseudoId(normalEmbedInIframeShadowRoot.firstChild) is "-webkit-plugin-replacement"
PASS normalEmbedInIframeShadowRoot.firstChild.firstChild.tagName is "IFRAME"
PASS successfullyParsed is true

TEST COMPLETE
